Data Mining CW1
Exercise:  
The assessment will be based on an essay type question.  The answers must be submitted 
electronically to Blackboard/Turnitin no later than the given date.  The work should all be 
clearly labelled with your details, including student number, and the details of the course.  
This piece of work is worth 35 % of the course marks. 
You must write a very concise essay of no more than 1,500 words on the topic of one or 
more main challenge/s remaining in data mining in 2024 that would be most beneficial to 
solve and explain what would be the impact of solving it on society or other stakeholders. In 
other words, what are the main challenges regarding the analysis of data in 2024,  or the 
use of such analysis in the real world and what would be the impact of solving those 
challenges.  You should perform a literature survey of work in this area and include all 
relevant references in your report.  The literature survey should be as up-to-date as possible 
(e.g. focused on last 5 years or less), so looking for articles that reflect the state-of-the art in 
the last few years.  You could cover any of the following issues: 
• what are some of the outstanding ethical challenges associated with data mining?; 
• what are the main algorithmic challenges for data mining?; 
• what are some of the main data challenges?; 
• what are some of the main application challenges? 
There may be other areas that come to light in your survey of the literature and those would 
be equally valid to cover but please focus on answering the question.  Also, you need to 
provide evidence by quoting peer reviewed literature that the challenges have been 
exposed/discussed by others.